+++ packages/analytics/analytics-chart/src/components/chart-plugins/DragSelectPlugin.ts
@@ -17,12 +17,14 @@ const dispatchEvent = (eventName: string, chart: Chart, pluginInstance: DragSele
   const xStartValue = chart.scales.x.getValueForPixel(pluginInstance.startX)
   const xEndValue = chart.scales.x.getValueForPixel(pluginInstance.endX)
 
-  chart.canvas.dispatchEvent(new CustomEvent<DragSelectEventDetail>(eventName, {
-    detail: {
-      xStart: xStartValue,
-      xEnd: xEndValue,
-    },
-  }))
+  if (xStartValue && xEndValue) {
+    chart.canvas.dispatchEvent(new CustomEvent<DragSelectEventDetail>(eventName, {
+      detail: {
+        xStart: Math.min(xStartValue, xEndValue),
+        xEnd: Math.max(xStartValue, xEndValue),
+      },
+    }))
+  }
 }
 
 export class DragSelectPlugin implements Plugin {