+++ packages/entities/entities-plugins/src/components/free-form/Datakit/visual-editor/constants.ts
@@ -1,0 +1,3 @@
+export const DK_HEADER_HEIGHT = '44px'
+export const DK_SIDE_PANEL_WIDTH = '220px'
+export const DK_NODE_PROPERTIES_PANEL_WIDTH = '366px'