+++ packages/entities/entities-plugins/src/components/free-form/Datakit/visual-editor/modal/EditorModal.vue
@@ -20,6 +20,7 @@
 import { useScrollLock } from '@vueuse/core'
 import { useFocusTrap } from '@vueuse/integrations/useFocusTrap'
 import { nextTick, useTemplateRef, watch } from 'vue'
+import { DK_HEADER_HEIGHT, DK_SIDE_PANEL_WIDTH } from '../constants'
 
 import EditorNav from './EditorNav.vue'
 import EditorContent from './EditorContent.vue'
@@ -71,8 +72,8 @@ function close() {
   z-index: 1000;
 
   /* stylelint-disable custom-property-pattern */
-  --dk-side-panel-width: 220px;
-  --dk-header-height: 44px;
+  --dk-side-panel-width: v-bind('DK_SIDE_PANEL_WIDTH');
+  --dk-header-height: v-bind('DK_HEADER_HEIGHT');
   /* stylelint-enable custom-property-pattern */
 
   .nav {